:recycle: Move init state logic to RNNLM
