+++ diagnnose/models/__init__.py
@@ -1,3 +1,2 @@
 from .import_model import import_model
-from .init_states import set_init_states
 from .language_model import LanguageModel
